Move not found auditing into exceptions
